--- txpool/block_service.go
@@ -1,5 +1,0 @@
-package txpool
-
-type BlockService interface {
-	ProposeBlock(transactions []Transaction) error
-}
