+++ txpool/infra/adapter/message_service.go
@@ -1,4 +1,4 @@
-package infra
+package adapter
 
 import (
 	"errors"
