+++ txpool/service.go
@@ -3,3 +3,7 @@ package txpool
 type MessageService interface {
 	SendLeaderTransactions(transactions []*Transaction, leader Leader) error
 }
+
+type BlockService interface {
+	ProposeBlock(transactions []Transaction) error
+}
